requirements.txt
requests
pyyaml
pandas
pydantic
schedule
python-dateutil
